milestone_4_task/requirements.txt
streamlit
pandas
numpy
altair
PyPDF2
python-docx
scikit-learn